Skybox tweaks
-Fixed destructor missing
-Changed skybox texture and renamed shader
-Removed default skybox, component still to do

+++ ScrapEngine/SimpleGame/GameObjects/Camera/GameCamera.cpp
@@ -16,7 +16,7 @@ void GameCamera::GameUpdate(float time)
 {
 	ScrapEngine::MouseLocation mouse = InputManagerRef->getLastMouseLocation();
 	
-	GameCameraRef->ProcessMouseMovement(mouse.xpos, mouse.ypos, true);
+	GameCameraRef->ProcessMouseMovement((float)mouse.xpos, (float)mouse.ypos, true);
 
 	cameraSpeed = 10.f * time;
 	if (InputManagerRef->getKeyboardKeyStatus(keyboard_key_W) == ScrapEngine::KeyboardKeyState::pressed) {
